requirements.txt
# Raspberry Pi: pip3 install -r requirements.txt
bme680>=1.0.5
flask>=3.0.0
PyYAML>=6.0.1
gpiozero>=2.0
